Optimize predictors in `lossless_transform` (#152)
Rewrite predictor transform 1 to be autovectorized.
Add an average function which performs average with bitwise operations.
LLVM is capable of doing this optimization but it's not able to
automatically perform it for every transform we have.
Add asserts to remove some saturating arithmetic instructions.

Predictors 10-13 were much slower when rewritten following
this pattern.